+++ src/encoding/evm/strategy_encoder/mod.rs
@@ -1,2 +1,2 @@
-mod encoder;
-mod selector;
+mod strategy_encoders;
+mod strategy_selector;
